Renamed PlayerState to PlayerView. Made state, history fields non-final so that they can be changed between steps in the game. Added numsFarmsBuilt and numFarmsLimit to add a cap to the number of farms built. Added a second constructor so that only the playerNum of the Agent needs specifying. This also allows PlayView to be instantiated upon Agent instantiation.
Added findBestResource(), which also considers the amount resource at the closest resources. Added boolean functions to test of a Unit has enough wood and gold to build a unit.

buildFarm(), depositResources(), buildPeasant(), and gatherBalancedResources() encapsulate the logic of these action creations.

isIncomplete() and getUnoccupiedUnits() allow Actions to assigned only to those Units who are not still executing a previously assigned Action.

Numerous getters and setters to wrap the identical methods for a StateView.

Added a few more Units enums: BARRACKS, FARM, and FOOTMAN.

Signed-off-by: rdtatton <rdt46@case.edu>