"Best" working version for submission.
As of now, this is the "best" working version using the added architecture from the `action` and 'playerView` package. However, some units (farms, barracks, and footmen) are not being built. This is peculiar because peasants are still being built and resources are being collected. Still have to debug more to find out what is causing this problem.

Signed-off-by: rdtatton <rdt46@case.edu>