"You want abstraction? You can't handle abstraction!"
Criterion and Goal are back in action! The type parameter of Goal is simply the objective type now, so that makes instantiating less messy. Also, got to practice using the Builder design pattern for Criterion and Goal which makes it look super clean in ResourceTracker.

UnitTracker and ResourceTracker are extensions of the general Tracker class which is ultimately just a Map whose semantics is meant to define what each status a tracked item has. In this case, Units have action statuses and resources have amount remaining.

Now, to finally revamp the rest of GameState.

Signed-off-by: rdt46@case.edu <rdt46@case.edu>