More GameState progress.
Beginning to clean up the Strips logic in GameState in terms of determining what all actions can be made and the resulting update of UnitTracker and ResourceTracker. Will need to double check the logic holds up once I'm done with possibleDirectives. In order to let the code do more of thinking, I am not specifying an ordering of the Resources. Doing so both simplifies the code logic (and abstracts it more, which is nice) and gets more at the notion of letting A star do its thing in finding the optimal plan.

Signed-off-by: rdt46@case.edu <rdt46@case.edu>