StripsAction 2.0.
For the StripsAction interface, I added a few methods: (1) effects() returns an EnumSet<StripsEnum> that define what a Unit status could be set to following this action; (2) computeCost() returns the cost to execute the general StripsAction; and (3) getStripsActionType() which connects the more complete StripsAction that contains enough state information to create a SEPIA Action with the StripsEnum that is used in GameState for high-level tracking.

Deposit, Gather, and Produce are the currently defined StripsActions. Probably won't get to Build, but we'll see. The general pattern is that preconditionsMet() checks to see if the StripsEnum exists in the current GameState UnitTracker; apply() uses the corresponding method in GameState to update the UnitTracker and ResourceTracker accordingly.

We'll see if this framework ends up working. Very similar to P4 with DirectiveMap, but a little less coupled. Each StripsAction is meant for a single Action. See StripsMap for details on how I plan to handle multiple actions for a single child of GameState.

Signed-off-by: rdt46@case.edu <rdt46@case.edu>