Done with StripsAction classes, for now.
Removed getStripsActionTyp() from the StripsAction interface since it does not make sense with StripsMap. StripsMap takes care of applying multiple StripsActions to the same GameState by looping through all the StripsMap actions, applying each to the same state. Because of how UnitTracker and ResourceTracker work with the actions, this should work, but we'll see.

Signed-off-by: rdt46@case.edu <rdt46@case.edu>