Minor updates to GameState.
Added toString() for Unit and Tracker.

Changed Criterion constructor to take a CriterionBuilder instance.

UnitTracker.build() sets Unit.initialAction to null if it is not set and adds it to the valid actions. Generally, initial action does not need to specified as a valid action as it is now automatically added to Unit.validActions.

Added an additional UnitTracker.build(int nCopies) that creates nCopies distinct instances of Units with the same valid and initial actions. This is a convenience method used when creating peasants.

Added Tracker.containsAny(S...s) to check if any variable number of s are present in the keys of the Tracker. This is used to check if one or more StripsEnums are present in a UnitTracker (see any of the classes that implement StripsAction).

Signed-off-by: rdt46@case.edu <rdt46@case.edu>