A Push in Defeat.
Well, I don't think I'm going to do very well on this project. The extent that I got is the initial GameState generation. No plan. Frankly, just spent too much time on GameState / didn't have enough time to debug. I think if I would've had another 10 hours to work on this (to debug), I could've finished it.

It's very disappointing. The core of my logic is there but there are still some aspects that need to be fixed. Currently, there is the infamous reference issue. When I debug and check to see the initial set of GameStates, they are all the same. Not sure why this is happening though as I spent time making copy constructors in order to peform a deep copy on all the inner classes.

Signed-off-by: rdt46@case.edu <rdt46@case.edu>